% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\docType{data}
\name{BETA_DRG_SCREEN_NAMES}
\alias{BETA_DRG_SCREEN_NAMES}
\title{The 24 MS-DRG screens, in the order the beta analysis program resolves them}
\format{
An object of class \code{character} of length 24.
}
\usage{
BETA_DRG_SCREEN_NAMES
}
\description{
Names are the SAS format names. \code{.beta_comorbidity_flags()} addresses
them by dropping the \code{DRG} suffix, so \code{CARDDRG} is \code{f("CARD")}.
}
\keyword{internal}
