% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\docType{data}
\name{BETA_FINAL_TARGETS}
\alias{BETA_FINAL_TARGETS}
\title{The 31 beta output columns}
\format{
An object of class \code{character} of length 31.
}
\usage{
BETA_FINAL_TARGETS
}
\description{
The 30 measures plus HTN_C, which the SAS derives last, after the MS-DRG
screen has run.
}
\keyword{internal}
