% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\docType{data}
\name{BETA_MEASURES}
\alias{BETA_MEASURES}
\title{The 30 beta comorbidity measures, in ARRAY COM1 order}
\format{
An object of class \code{character} of length 30.
}
\usage{
BETA_MEASURES
}
\description{
The 30 beta comorbidity measures, in ARRAY COM1 order
}
\keyword{internal}
