% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_builders.R
\name{build_poa_exempt_formats}
\alias{build_poa_exempt_formats}
\title{Build POA-exempt code sets from CSV file}
\usage{
build_poa_exempt_formats(poaxmpt_csv_path)
}
\arguments{
\item{poaxmpt_csv_path}{Character, path to CSV file with "version" and "code" columns}
}
\value{
Named list with POA exempt codes by version (e.g., list(v33 = c(...), v34 = c(...)))
}
\description{
Internal function that creates a named list of POA (Present on Admission) exempt 
codes organized by ICD version. This function is used internally by the 
comorbidity() function.
}
\keyword{internal}
