% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\docType{data}
\name{CMR_COMBO_TARGETS}
\alias{CMR_COMBO_TARGETS}
\title{Combination targets, which fan out into component flags}
\format{
An object of class \code{character} of length 10.
}
\usage{
CMR_COMBO_TARGETS
}
\description{
Combination targets, which fan out into component flags
}
\keyword{internal}
