% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\docType{data}
\name{CMR_POA_NEUTRAL}
\alias{CMR_POA_NEUTRAL}
\title{Targets assigned regardless of POA}
\format{
An object of class \code{character} of length 20.
}
\usage{
CMR_POA_NEUTRAL
}
\description{
Targets assigned regardless of POA
}
\keyword{internal}
