% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity.R
\name{.apply_comorbidity_rules}
\alias{.apply_comorbidity_rules}
\title{Internal function to apply comorbidity business rules}
\usage{
.apply_comorbidity_rules(dx_with_targets, n_rows, use_poa)
}
\description{
Internal function to apply comorbidity business rules
}
\keyword{internal}
