% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity.R
\name{.apply_exclusion_rules}
\alias{.apply_exclusion_rules}
\title{Apply hierarchical exclusion rules}
\usage{
.apply_exclusion_rules(result_matrix, use_poa, n_rows)
}
\description{
Apply hierarchical exclusion rules
}
\keyword{internal}
