% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity.R
\name{.build_result_matrix}
\alias{.build_result_matrix}
\title{Build the final result matrix with comorbidity flags}
\usage{
.build_result_matrix(all_assignments, n_rows, use_poa)
}
\description{
Build the final result matrix with comorbidity flags
}
\keyword{internal}
