% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity.R
\name{.handle_combination_codes}
\alias{.handle_combination_codes}
\title{Handle combination codes logic}
\usage{
.handle_combination_codes(dx_with_targets, use_poa)
}
\description{
Handle combination codes logic
}
\keyword{internal}
