% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity.R
\name{.handle_special_cases}
\alias{.handle_special_cases}
\title{Handle special cases and combination codes}
\usage{
.handle_special_cases(valid_assignments, dx_with_targets, use_poa)
}
\description{
Handle special cases and combination codes
}
\keyword{internal}
