% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\docType{data}
\name{CMR_POA_DEPENDENT}
\alias{CMR_POA_DEPENDENT}
\title{Targets assigned only when POA is absent, exempt, or "Y"/"W"}
\format{
An object of class \code{character} of length 19.
}
\usage{
CMR_POA_DEPENDENT
}
\description{
Targets assigned only when POA is absent, exempt, or "Y"/"W"
}
\keyword{internal}
