% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cmr_releases}
\alias{cmr_releases}
\title{AHRQ CMR releases supported by this package}
\usage{
cmr_releases()
}
\value{
Character vector of release labels, e.g. \code{c("2022.1", ...)}.
}
\description{
The AHRQ/HCUP "Elixhauser Comorbidity Software Refined for ICD-10-CM" releases
that \code{\link{comorbidity}} and \code{\link{cmr_index}} can score with,
oldest first. Any one of these is valid as their \code{release} argument.
}
\examples{
cmr_releases()
}
\seealso{
\code{\link{cmr_version}} for the default.
}
