% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_builders.R
\name{determine_icd_version}
\alias{determine_icd_version}
\title{Determine ICD version based on year and quarter}
\usage{
determine_icd_version(year, quarter)
}
\arguments{
\item{year}{Integer vector of years}

\item{quarter}{Integer vector of quarters (1-4)}
}
\value{
Integer vector of ICD-10-CM version numbers
}
\description{
Internal function that maps year and quarter combinations to ICD-10-CM version 
numbers following the standard CMS versioning timeline. This function is used 
internally by the comorbidity() function.
}
\keyword{internal}
